+++ src/antares/study/version/__about__.py
@@ -4,7 +4,7 @@ Antares Study (and Solver) version models.
 
 # Standard project metadata
 
-__version__ = "1.0.5"
+__version__ = "1.0.6"
 __author__ = "RTE, Antares Web Team"
 __date__ = "2024-07-06"
 __credits__ = "© Réseau de Transport de l’Électricité (RTE)"